+++ tasks.py
@@ -144,10 +144,10 @@ def update_doc(ctx):
     make_doc(ctx)
     try:
         contribute_dash(ctx)
+        ctx.run("mv pymatgen.tgz ..", warn=True)
     except:
         pass
     ctx.run("cp docs_rst/conf-normal.py docs_rst/conf.py")
-    ctx.run("mv pymatgen.tgz ..")
     make_doc(ctx)
     ctx.run("git add .")
     ctx.run("git commit -a -m \"Update docs\"")