+++ docs_rst/change_log.rst
@@ -1,6 +1,19 @@
 Change log
 ==========
 
+v2019.8.14
+----------
+* Update DOSCAR from lobster (@JaGEO)
+* PerturbStructureTransformation (@rees-c)
+* Misc bug fixes.
+
+v2019.7.30
+----------
+* Bug fixes (@shyuep, @mfherbst)
+* More type hint annotations (@shyuep)
+* Improvements to BabelMolAdaptor (@smheidrich)
+* Convenience Transformations for AdsorbateSiteFinder (@mkhorton)
+
 v2019.7.21
 ----------
 * Add CubicSupercellTransformation and PerturbedSupercellsTransformation (@rees-c, @utf)
@@ -9,6 +22,7 @@ v2019.7.21
 * Improved Lobster interface (@JaGeo)
 * Bug fixes (@sthartman, @dwinston, @utf)
 * New functionality for calculation of Heisenberg exchange parameters (@ncfrey)
+* Improvements to Miller indices handling and Lattice (@richardtran415)
 
 
 v2019.7.2